WHAT: Movement has been fixed to prevent the annoying diagonal movement of before.
HOW:
The movement now works on an input stack. When being held the input (UP, DOWN, LEFT, or RIGHT) are added to the stack (keeping the order of which was pressed first). Once released the input is removed from the list similarly.

Then the system goes through the stack and attempts to apply the inputs in order, once a valid input has been applied the loop breaks and no more inputs are applied.